Mohammed Salah 03-05-2022 09:50 PM

+++ app/Http/Controllers/Backend/Admin/CourcesController.php
@@ -195,17 +195,16 @@ class CourcesController extends Controller
 
             $path = storage_path('app/public/videos');
             $files = File::allFiles($path);
-            // if (count($files) > 0) {
-            //     foreach ($files as $file) {
-                // return $files[0];
+            if (count($files) > 0) {
+                foreach ($files as $file) {
                     $file_replaced = str_replace(storage_path('app').'/', '', $files[0]);
-                    return $file_replaced;
+                    // return $file_replaced;
                     $section_video = CourseSection::where('video',$file_replaced)->get()->first();
                     if(!$section_video){
                         File::delete($file_replaced);
                     }
-            //     }
-            // }
+                }
+            }
 
 
             $course = Course::find($id);
